Use our own logic to stop the Neo4j service in the deb package
We do this, rather than using start-stop-daemon, so that the stopping
behaviour is consistent between tarball and deb package. In particular
we want to make sure that the stop is always gentle rather than brutal,
so that operators can chose themselves whether to terminate the process
if it's very slow shutting down.

This means a small behavioural change in order to adhere to the standard
service behaviour: we no longer consider it an error to stop an
already-stopped service.