Fix a test configuration for one of the page cache tests
The `PageCacheSlowTest.mustNotLoseUpdatesWhenOpeningMultiplePageCursorsPerThread()` test was flaky and would randomly fail.
It had been carelessly configured such that the number of threads accessing the page cache, multiplied by the maximum number of page cursors they could open, was exceding the number of cache pages.
This meant that it was possible for the test to run into live locks, where a page fault would fruitlessly look for a page to cooperatively evict, only to discover that all pages in the cache were already locked for other purposes.
The test is now tuned such that this cannot occur - not even if the test repitition count is dramatically increased; not even if the live-lock threshold is dramatically decreased.