Correct voting logic.
"Raft determines which of two logs is more up-to-date by comparing
the index and term of the last entries in the logs.
If the logs have last entries with different terms, then the log with
the later term is more up-to-date.
If the logs end with the same term, then whichever log is longer is
more up-to-date."

Previously the code required the candidate log to be at least as long
as the follower's log, even if the candidate's log ended at a later
term. That was incorrect.