Inline RAFT message processing.
Process RAFT messages directly inside the role handlers, thereby removing
the need to change role, and having to send the message back for later
processing by that role.

The tests have been restructured to reflect the requirement for
some messages to be handled in mostly the same way, regardless of
the current role.