fix for having long strings in tx logs
This fixes an issue causing a NegativeArraySizeException when reading from a log containing an write operation of a string larger 32k to a manual index.
The cause for this is that conversion of short to int must be masked to prevent misbehaviour if the short value is negative (short is signed!).
Needed to enhance InMemoryLogChannel to support a configurable buffer size.