Temporarily allow the Desktop app to define the location of a database
We have just changed the way we specify a database from by-path to
by-name (see previous commit). But in order to the Desktop app to work
that way it also needs to be able to specify locations for config and
log directories. The work to support that hasn't been done yet. So we
temporarily allow the Desktop app to use the old approach and will
revert this commit once the other changes have been made and the Desktop
brought into line with the new way of doing things.