Support open options DELETE_ON_CLOSE and EXCLUSIVE for PageCache.map()
The `StandardOpenOption.DELETE_ON_CLOSE` will be useful for migrations.
Specifically, it has a use when migrating a store to a block device based storage system.
The `PageCacheOpenOption.EXCLUSIVE` may find use in the counts store, and generally in places where we want to ensure a single point of access to a file.