Specify sender address for network communication
This is a necessary change to make sure we transmit data on the right
path to the destination. Before this, neo4j-ha would not work in case
there were multiple IP addresses on each network interface, since "the
first address" would be picked as the transmission origin.

Also adds more descriptive debug logging around connections.