Add progress monitor to DirectRecordStoreMigrator
DirectRecordStoreMigrator is used to migrate between versions that have same
store format capabilities. It migrates records for corresponding stores
sequentially.

Currently it only prints progress in one go when the migration is completed.

This commit adds progress tracking so that DirectRecordStoreMigrator outputs
dots while doing it's job. Percentage is calculated based on the numbers of
migrated stores. It would me more precise to do this based on sum of all
highIds but there is a high risk of long overflow.