Process all property updates from particular node as single update during index population
Group all updates from all properties on single node into container that will be processed
in atomic fashion - all of the updates will be processed before check of population queue for current node.
All updates need to be processed as single unit to avoid duplicates in index that can appear
when queue already contains update for current node, but for some other property.