Fix reading records via RecordCursor
Currently record cursor has an optimization where it does not reposition the
underlying page cursor if it points to the correct page id. However this is
not entirely correct because the underlying page might have been evicted and
we might attempt reading garbage data. This resulted in failures during
reading properties.

This commit removes page id check.