(MAINT) Update powershell scripts to emit exit codes instead of exit
It was found that running the powershell module would prematurely exit
the shell e.g. Running Invoke-Neo4j would exit the current powershell
session.  This was due to the Exit statements being in the powershell
helpers.  This commit moves the Exit statements to the helper BAT files
and changes the powershell entry functions to return an exit code