First draft at making the http.log use our own logging infrastucture.
This currently fails a number of tests, because they still assume that they can configure the http.log through the logback xml file.
So while the actual logging part is more or less done, the configuration part is still up in the air.
Another thing that needs to be looked at, is the http log line format.
I've tried my best to make our logging framework print lines that look like the logback pattern, but I'm not sure if its good enough yet.