Make the tests for renewable timeout service unsensitive to timing.
These changes attempt to better catch time/event order in the tests and
for example not assert on things to happen in time on anything less than
a very long time basis, e.g. 30 seconds. So the tests now will only test
that things do happen after a certain minimum time or that things did
not happen up to a maximum time.

The base value of the timer used in tests has also been made shorter
so that the test in non-failing cases finishes quickly.