Replace Windows Service Wrapper with ACD ProcRun
The Java Windows Service Wrapper has been replaced, and instead
the Apache Commons Daemon ProcRun is used for running the Neo4j
Server as a Windows Service.  This commit modifies the powershell
scripts to use PROCSRV instead.  It supports 32bit and 64bit (AMD)
architecture (Not ia64)

http://commons.apache.org/proper/commons-daemon/procrun.html

The ProcRun application can only run a registered Windows Service, not
an adhoc console session.  The Get-Java function was modified to call
Java directly when starting Neo4j as a console session.