More aggressive inconsistent reads from the PageCacheRule
We previously just returned zeros from the adversarial page read cursors when doing an inconsistent read.
Now, we count the number of operations performed against the page cursor, choose one, tell the reader to retry, and then we return random garbage data from the point of the chosed operation.
This finds many, many more cases of faulty use of the page cursors.