Fix some ElectionPerformanceIT issues.
If a timeout exception occured then a disconnected member
would not be reconnected. Now we always reset the complete test
network to a pristine state after each iteration.

We now also wait for the initial member set to be fully propagated
before starting the scenario, so that all members are electable.