Make the inconsistency injection of the `PageCacheRule` more devious
Previously, the `PageCacheRule` would by default randomly decide that some reads would be inconsistent.
Then, for those inconsistent reads, the page cursor would return zeros for all read operations.

Now, we still randomly choose in the `PageCursor.next()` call if a read will be inconsistent, but instead of just returning zeros, we are now more devious:
To start with, we just delegate all read operations and let them succeed, counting them in the process.
Then, in the `shouldRetry` method, we choose an operation at random to be the start of the inconsistent read.
We then let the reader retry, and when they get to the victim operation, we start returning random data from that point on.

This surfaces a number of bugs, which have also been fixed in this commit.