Diverge physical raft log from shared kernel implementation.
PhysicalLogFile and friends don't work well with the need to truncate
the log file, as required for a raft log implementation.

In this commit we duplicate some classes into the core-edge module,
but this is only an intermediary step before we simplify the classes
on the core-edge side. We also temporarily increase visibility of
some kernel classes to make the move possible.

This commit also remove storing commit index from the responsibility
of raft log implementations, since we now honour the literature and treat
commit index as volatile state.