Address review issues of adversarial and devious pages cursors
This also fixes a bug in `NodeStore.loadRecord` that was affecting the PBI, where an inconsistent read would make a node record look like it was in use, and the loadRecord would fill it with garbage.
However, when the record decoding retried after the inconsistent read, and the record was discovered to NOT be in use, the garbage data and raised inUse flag would be left in the NodeRecord object.
The `NodeCountsProcessor` in PBI would then blow up on the left-over garbage data.