Don't default auth store location to a path that doesn't make sense in the kernel
The old default assumed the existence of a data directory, which doesn't
make sense in the kernel. This changes it so that there is no default
and a value has to be provided if authentication is turned on.

The server always provides a value (and will override any value that
external config tries to set), giving the desired path of
data/dbms/auth.

This means that the Desktop can no longer set a value for this
explicitly, so we set a location for the data directory instead.