Fix variable exposure through multiple projections
Variables introduced before multiple projections in update queries would
not be exposed after the projections that introduced empty query graph tails.
This would cause bugs with e.g. `CREATE` creating nodes that were already bound.

The fix is to add all currently exposed symbols as arguments to the
new empty query graph when adding a new tail to the PlannerQueryBuilder.