Correctly identify store files when deleting them
When a slave joins an HA cluster it may need to delete its store
files (because they represent a different logical database). The code to
do that is pretty brutal -- it deletes everything in the store dir apart
from some blacklisted files. Now that we keep the logs and other
directories in the store dir when running embedded, we need to update
the blacklist.

This problem was failing the build on Windows (because file locks are
exclusive on that platform), but there was a bug for any platform
running embedded HA.