Fix how the check on the file page size in PageCache.map interacts with PageCacheOpenOptions.ANY_PAGE_SIZE
Our block device integration needs to map files without specifying a page size.
Recent commits added a check which required that the file page size be at least 8 bytes long, but this unfortunately broke our block device integration code.
This changes the behaviour of this check, to only apply when we actually need the file page size to make sense.